// src/app/page/master/map-kelompokevaluasi-jabatan/map-kelompokevaluasi-jabatan.interface.ts
import {EventEmitter} from '@angular/core';

export interface MapKelompokevaluasiJabatan {
	kdProduk?: string;
	kdEvent?: number;

	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	label?: any;
	value?: any;
}